Major: USERS Api complete 1.) (Add, Get, Update, Delete) All operations working 2.) Refactoring, Created /Services and Inside that created user-management-serv.js for userService 3.) All operations related to users are defined here
