--- .data/users/7022420170.json
@@ -1,1 +1,0 @@
-{"firstName":"Ravi","lastName":"Upadhyay","phone":"7022420170","hashedPassword":"4da9875ab3e6c31cf2710f4eb72f01326fc64a542f99d215924acf4094d58296","tosAgreement":true}
