--- .data/users/9460616932.json
@@ -1,1 +1,0 @@
-{"firstName":"Madhu","lastName":"Panecha","phone":"9460616932","hashedPassword":"4da9875ab3e6c31cf2710f4eb72f01326fc64a542f99d215924acf4094d58296","tosAgreement":true}
