+++ lib/data-handler-lib.js
@@ -100,7 +100,7 @@ dataHandler.delete = (dir, file, callback) => {
     const completePath = `${dataHandler.baseDataDir}${dir}/${file}.json`;
     fs.unlink(completePath, (errorD) => {
         if(!errorD) {
-
+            callback(false);
         } else {
             callback(errors.couldNotDel);
         }
